[core|gym] Log data using 64bits variables, Add unit tests to Gym Jiminy. (#222)
* [misc] Remove old deprecated docstring style.
* [misc] Enable to re-run automatic build scripts.
* [misc] Install numba on CI instead of 'easy_install_ubuntu.sh' script.
* [misc] Add-back Pinocchio StdVec_* converters.
* [misc] Fix Pinocchio imported twise that was messing up with boost python converters on Windows.
* [core] Log variable in 64bits instead of 32bits (and drop compatibility with old proprietary parser).
* [core] rename constant 'TELEMETRY_MAX_BUFFER_SIZE' in 'TELEMETRY_MIN_BUFFER_SIZE' for consistency.
* [core] Consistent style over the whole codebase.
* [core] Do not raise an exception if qhull not available and warn only.
* [core|python] Fix unit tests for new logging precision.
* [python] Use 'tempfile.mkstemp' instead of 'tempfile.NamedTemporaryFile' to create temp files (Windows compatibility).
* [python] Fix deprecation warnings.
* [python/meshcat] Fix irrelevant warning and error messages.
* [python|gym] Add flag to enable/disable verbose logging related to robot toml file handling.
* [python|gym] Various bug fixes. Improve support of Windows.
* [gym] Always register controller target for internal computation.
* [gym] Goes back to float64 for consistency with Jiminy float type.
* [gym] add 'get_log' helper at 'BaseJiminyEnv' level.
* [gym] 'register_variables' now returns 'is_success' flag.
* [gym] Add unit tests.
* [gym] Use 'meshcat' backend by default if rendering with mode 'rgb_array'.

+++ core/include/jiminy/core/telemetry/TelemetryRecorder.h
@@ -62,8 +62,8 @@ namespace jiminy
         hresult_t writeDataBinary(std::string const & filename);
         static void getData(std::vector<std::string>                   & header,
                             std::vector<float64_t>                     & timestamps,
-                            std::vector<std::vector<int32_t> >         & intData,
-                            std::vector<std::vector<float32_t> >       & floatData,
+                            std::vector<std::vector<int64_t> >         & intData,
+                            std::vector<std::vector<float64_t> >       & floatData,
                             std::vector<AbstractIODevice *>            & flows,
                             int64_t                              const & integerSectionSize,
                             int64_t                              const & floatSectionSize,
@@ -71,8 +71,8 @@ namespace jiminy
                             int64_t                                      recordedBytesDataLine = -1);
         void getData(std::vector<std::string>             & header,
                      std::vector<float64_t>               & timestamps,
-                     std::vector<std::vector<int32_t> >   & intData,
-                     std::vector<std::vector<float32_t> > & floatData);
+                     std::vector<std::vector<int64_t> >   & intData,
+                     std::vector<std::vector<float64_t> > & floatData);
     private:
         ////////////////////////////////////////////////////////////////////////
         /// \brief   Create a new file to continue the recording.
